Extract 2 stream sources: file and ytdlp
