+++ src/player/symphonia/controller/ytdlp_source.rs
@@ -1,0 +1,7 @@
+use symphonia::core::{audio::AudioBufferRef, codecs::Decoder};
+
+type Result<T> = std::result::Result<T, Box<dyn std::error::Error>>;
+
+pub(super) fn start_player_thread(url: &str) {
+    todo!()
+}
